src/main.py
#!/usr/bin/env python3
"""
Automated Video Generator - Main Entry Point
"""

import click
import logging
from pathlib import Path

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)


@click.command()
@click.option('--input', '-i', help='Input directory or file path')
@click.option('--output', '-o', help='Output directory path')
@click.option('--config', '-c', help='Configuration file path')
@click.option('--verbose', '-v', is_flag=True, help='Enable verbose logging')
def main(input, output, config, verbose):
    """Automated Video Generator - Generate and upload videos to platforms."""
    
    if verbose:
        logging.getLogger().setLevel(logging.DEBUG)
    
    logger.info("Starting Automated Video Generator")
    
    # TODO: Implement video generation logic
    click.echo("Video generator is ready! Implementation coming soon...")
    
    if input:
        click.echo(f"Input: {input}")
    if output:
        click.echo(f"Output: {output}")
    if config:
        click.echo(f"Config: {config}")


if __name__ == '__main__':
    main()
