requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Task scheduling
apscheduler>=3.10.0

# Video processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0
pillow>=10.0.0

# Image generation
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
torch>=2.0.0
torchvision>=0.15.0
safetensors>=0.3.0
huggingface-hub>=0.16.0

# API and HTTP
requests>=2.31.0
flask>=2.3.0

# Pyramid Flow and Gradio
gradio-client>=1.11.0
httpx>=0.24.0
websockets>=10.0

# YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth>=2.22.0

# Google Sheets
gspread>=5.11.0

# Text-to-speech
TTS>=0.20.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
pathlib2>=2.3.0
psutil>=5.9.0
pyyaml>=6.0
zipp>=3.15.0
hf-xet>=0.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
